Store actions losslessly so replay verification is exact
Replay verification failed at 1e-3 deviation. Diagnosis, rather than a looser
tolerance: replaying float64 actions reproduces the trajectory to 0.000e+00, so
the simulator is exactly deterministic. Replaying the same actions quantized to
float32 deviates 1.87e-04 -- from a quantization error of 2.95e-08 rad, which
the dynamics amplify roughly 6300x over 17 steps.

So the defect was lossy storage, not nondeterminism. Actions are now stored
float64; next to 224x224x3 images they are a rounding error in the byte budget.
Replay tolerance tightened 1e-4 -> 1e-6 accordingly, since replay should now be
exact rather than merely close.

Worth carrying forward to M3: a 6300x amplification over 17 steps says this
system diverges fast from tiny perturbations, which bounds how far an open-loop
latent rollout can be trusted before it stops describing the same trajectory.

Also lowers the IK gain 0.12 -> 0.07; 15-frame episodes barely contain a
trajectory to imitate.

+++ src/jetspace/data/episode.py
@@ -57,7 +57,12 @@ class EpisodeBuffer:
         for cam, frame in pixels.items():
             self.pixels.setdefault(cam, []).append(np.asarray(frame, dtype=np.uint8))
         self.proprio.append(np.asarray(proprio, dtype=np.float32))
-        self.action.append(np.asarray(action, dtype=np.float32))
+        # Actions are stored float64, deliberately. Quantizing to float32 costs
+        # ~3e-08 rad, which the dynamics amplify ~6300x over a 17-step episode
+        # to ~2e-04 rad -- enough to break exact replay verification. Actions are
+        # a rounding error in the byte budget next to 224x224x3 images, so there
+        # is no reason to lose the precision.
+        self.action.append(np.asarray(action, dtype=np.float64))
         self.reward.append(float(reward))
         self.success.append(bool(success))
